Restructured core and changed build + zip commands accordingly
