+++ .gitignore
@@ -166,9 +166,9 @@ dmypy.json
 *.pch
 
 # Compiled Dynamic libraries
-*.so
-*.dylib
-*.dll
+**/*.so
+**/*.dylib
+**/*.dll
 
 # Fortran module files
 *.mod
