+++ prospr/core/src/depth_first.cpp
