+++ prospr/core/src/depth_first_bnb.cpp
