+++ prospr/core/src/protein.hpp
